薪资模块的更新，薪资标准、奖惩模块的增删查改

+++ src/main/java/com/tomorrow/entity/Salary.java
@@ -52,7 +52,7 @@ public class Salary {
     private double late;
 
     @TableField(value = "absenteeism")
-    private double absenteeism;
+    private int absenteeism;
 
     @TableField(value = "actual")
     private double actual;
